delete older experiment code
remove modules related to the initial attempt Algorithm and
Hyperparameter controllers, combined to make the MLFController.
These models are defunct because the state space to search was
too large, i.e. the output space consisted of some `n` sequence
of algorithms, where the validity was accounted for by a positive
reward. Not a good direction to go.

fyi: the updated `cash_controller` + `reinforce` algorithm encodes
the structure of valid algorithm combinations