move backward pass into CASHReinforce, fix bug
fixes #1. this diff moves `backward` method into CASHReinforce,
where it belongs! this improves overall modularity of the system,
since the model architecture is agnostic of the RL algorithm
training it. it also catches a bug where the baseline reward
buffer is not trimmed to only use the most recent baseline
rewards, so the model was working off of the values from the
first episode... this might explain why models degrade in
performance after a while.