improve the way error logging is done
the TaskEnvironment module is where this is really important.
this is because the `pynisher` dependency that it uses to
enforce memory and runtime limits catches all exceptions that
occur during MLF fitting and prediction as `pynisher.AnythingException`s.
Probably not the ideal behavior for deep cash, but until I
can work on my own implementation for enforcing memory and runtime
limits, I'll have to work with `pynisher`.

Now the deep cash system's logging works as intended. There's a
`utils.init_logging` function that takes a `log_path` argument
so that all fitting and prediction errors are written to that
file.

The next step would be to decide whether or not to return INVALID
fit/predict calls in the `_ml_framework_fitter` and
`_ml_framework_predict` functions so that the `TaskEnvironment.fit`
methods can actually error out when there's an INVALID exception.
This would improve the reliability of the system, since right now
all errors happily return None, and the CASH controller perceives
this is a negative reward.