add Scorer named tuple to include scorer metadata
this diff defines a `Scorer` namedtuple that holds more metadata
about a scorer: a reward transformation function and a score
comparator to determine if one score is better than another.
This is implemented so that the `best_validation_score` metadata
is correct for loss function scorers like mean_squared_error.