add openml regression datasets
this diff adds a DataSourceType enum to data_types and
creates a parser function for the openml api. this function
currently parses regression tasks and obtains the source
dataset, eventually this should also apply to the classification
function.

For now, just to maintain consistency in the classification
environment, I'll keep the classification dataset parser
as is.