add kaggle regression and classification exps
this diff adds experiment configs that test the cash
controller's abilities on the kaggle datasets. The
resulting controllers after training should be
evaluated on the test set performance. This is a naive
run because the cash controller validation metric may
not be the same as the kaggle competition performance
metric. This problem will be rectified in a future
set of experiments.