train agent only on kaggle poker rule induction
in experiment `kaggle_classification_exp_1b`, the poker
induction mean reward didn't exceed `0.5`. The hypothesis
behind this is that the agent learns to proposes MLFs that
are better for the task environments that is has an easier
time doing well in.

Testing this hypothesis by training an agent exclusively
on the poker rule induction dataset, expecting to get
higher performance.