implement inference engine using trained controller
this diff creates a new object CASHInference as the prediction
engine of the CASH controller. A controller can be evaluated
in two ways:

1. evaluating the test set performonce in the training data
   environment distribution.
2. evaluating the validation/test set performance in the
   test data environment distribution.

Made several key updates in the data environment API that enables
better interface for accessing data. There's still a long way
to go to make it better... right now too much state is being
stored in the TaskEnvironment object. Makes modular design
more challenging. Need to not rely too much on state.