action_{modules, meta, index} keeps track of actions
this diff implements a big change to the way that actions are
indexed in the `MetaLearnController` class. Before, it was
difficult to keep track of and extend the action space to new
sets of algorithm components and hyperparameters since module
metadata was stored in a list. Now that the modules are stored
in a `nn.ModuleDict`, a pre-trained controller can be extended
with additional algorithms and hyperparameters without the risk
of overriding pre-trained action weights, as long as the
algorithm components and hyperparameters are named uniquely.